[ios] Add Xcode 11.6.0, 12.0 to nightly jobs. (#349)

+++ platform/ios/test/MGLNetworkConfigurationTests.m
@@ -21,7 +21,7 @@
     
     for (NSUInteger i = 0; i < numberOfConcurrentBlocks; i++) {
         
-        NSString *event = [NSString stringWithFormat:@"test://event-%ld", i];
+        NSString *event = [NSString stringWithFormat:@"test://event-%lu", (unsigned long)i];
         NSString *resourceType = @"test";
         
         dispatch_async(queue, ^{
